Seperate wallet and main test
