+++ testing/wasm/src/testData.json
@@ -7,6 +7,7 @@
         "prefix": "kipsi:qr36zdxs0dn3n0h799jhdl02qks5742lxjgmfsfj9xmlca7n4l6mw0s0n48nx",
         "content": "kaspa:asdfasdfkaspaasdf"
       }
-    }
+    },
+    "mnemonics": "blue fantasy sweet embody prepare mutual drum glide useful stool nice tell observe woman cannon pulp found wonder treat police blossom federal sure disagree"
   }
 }
